add env setup
clean up
