+++ .gitignore
@@ -13,5 +13,7 @@
 .vscode/
 .DS_Store
 **/.ipynb_checkpoints/*
+**/__pycache__/*
 **/data*/
 **/plots*
+
