+++ Electrical_Panels_EUSS/__version__.py
@@ -1,0 +1,9 @@
+import datetime as dt
+
+__title__ = 'panels'
+__name__ = 'panels'
+__description__ = 'Estimation of Electrical Panel Ampacity for ResStock buildings'
+__url__ = 'https://github.com/NREL/resstock/tree/electrical_panels'
+__version__ = '0.1.0'
+__license__ = 'BSD-3'
+__copyright__ = 'Copyright {} The Alliance for Sustainable Energy'.format(dt.date.today().year)
