+++ Electrical_Panels_EUSS/panels.egg-info/PKG-INFO
@@ -1,0 +1,7 @@
+Metadata-Version: 2.1
+Name: panels
+Version: 0.1.0
+Summary: Estimation of Electrical Panel Ampacity for ResStock buildings
+Home-page: https://github.com/NREL/resstock/tree/electrical_panels
+Requires-Python: >=3.10
+Provides-Extra: dev
