+++ Electrical_Panels_EUSS/panels.egg-info/SOURCES.txt
@@ -1,0 +1,7 @@
+README.md
+setup.py
+panels.egg-info/PKG-INFO
+panels.egg-info/SOURCES.txt
+panels.egg-info/dependency_links.txt
+panels.egg-info/requires.txt
+panels.egg-info/top_level.txt
