+++ Electrical_Panels_EUSS/panels.egg-info/requires.txt
@@ -1,0 +1,15 @@
+pandas<2
+ipython
+pyarrow
+scipy
+openpyxl
+matplotlib
+scikit-learn==0.24.2
+
+[dev]
+pytest
+flake8==3.8.2
+boto3
+awscli
+black
+plotly
